+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.2.10 patch
 * Fix [#170](https://github.com/mei-friend/mei-friend/issues/170) to load both types of quotes for MEI schema information
+* Show breaks option name as tooltip (title) in breaks select (https://github.com/rism-digital/verovio/issues/4196)
 
 ### 1.2.9 patch
 * Fix auto indentation when adding a zone to facsimile